# Optimize mask memory for flash attention